+++ R/val_prep_pipeline.R
@@ -83,13 +83,15 @@ val_prep_pipeline <- function(
   metric_pkg <- match.arg(metric_pkg)
   stopifnot(inherits(as.Date(val_date), c("Date", "POSIXt")))
   apply_verbose(verbose)
-  configure_bioc_repositories_if_requested(quiet = TRUE)
-  configure_riskmetric_offline_if_requested(quiet = TRUE)
 
   # Route pull_config() at any depth to the user-supplied config, if any.
   old_cfg <- options()["val.pipeline.config_path"]
   on.exit(options(old_cfg), add = TRUE)
   apply_config_path(config_path)
+
+  validate_config()
+  configure_bioc_repositories_if_requested(quiet = TRUE)
+  configure_riskmetric_offline_if_requested(quiet = TRUE)
 
   #
   # ---- Set time variables ----